Able to place comments on products in a order
